fix: added the correct activity vacancies visualization
