+++ src/components/Activities/List.js
@@ -1,9 +1,6 @@
 import styled from 'styled-components';
 import dayjs from 'dayjs';
-import { CgEnter, CgCheckO, CgCloseO } from 'react-icons/cg';
-import Swal from 'sweetalert2';
-import { toast } from 'react-toastify';
-import useCreateActivity from '../../hooks/api/useCreateActivity';
+import CapacityBox from './CapacityBox';
 
 function filterActivitiesByLocalization({ activities, localizationId }) {
   const filteredActivities = activities?.filter((activity) => activity.Localization.id === localizationId);
@@ -21,32 +18,9 @@ function differenceBetweenDatesInHours({ startTime, endTime }) {
   return difference;
 }
 
-export default function List({ activitiesFromDay, localizationId }) {
+export default function List({ activitiesFromDay, localizationId, selectedDate }) {
   const filteredActivities = filterActivitiesByLocalization({ activities: activitiesFromDay, localizationId });
   sortActivitiesByTime(filteredActivities);
-  const { createActivity } = useCreateActivity();
-
-  function bookActivity(activityId) {
-    Swal.fire({
-      title: 'Tem certeza que deseja se insecrever?',
-      text: 'Você não poderá retirar a inscrição depois!',
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#3085d6',
-      cancelButtonColor: '#d33',
-      confirmButtonText: 'Confirmar!',
-      cancelButtonText: 'Cancelar',
-    }).then(async(result) => {
-      if (result.isConfirmed) {
-        try {
-          await createActivity({ activityId });
-          toast('Inscrição confirmada!');
-        } catch (error) {
-          toast('A incrição falou, você possui outra atividade neste horário!');
-        }
-      }
-    });
-  }
 
   return (
     <Wrapper>
@@ -58,23 +32,7 @@ export default function List({ activitiesFromDay, localizationId }) {
               {activity.startTime.slice(11, 16)} - {activity.endTime.slice(11, 16)}
             </div>
           </div>
-          <CapacityBox capacity={activity.capacity}>
-            {activity.capacity > 0 ? (
-              <div onClick={() => bookActivity(activity.id)}>
-                <div>
-                  <CgEnter color="#078632" size="22px" />
-                </div>
-                <div>{activity.capacity} vagas</div>
-              </div>
-            ) : (
-              <>
-                <div>
-                  <CgCloseO color="#CC6666" size="18px" />
-                </div>
-                <div>Esgotado</div>
-              </>
-            )}
-          </CapacityBox>
+          <CapacityBox key={index} activity={activity} selectedDate={selectedDate} />
         </Lecture>
       ))}
     </Wrapper>
@@ -125,23 +83,5 @@ const Lecture = styled.div`
     flex-direction: column;
     justify-content: center;
     align-items: center;
-  }
-`;
-
-const CapacityBox = styled.div`
-  font-weight: 400;
-  font-size: 9px;
-  line-height: 11px;
-  padding-left: 8px;
-  color: ${(props) => (props.capacity > 0 ? '#078632' : '#CC6666')};
-  cursor: ${(props) => (props.capacity > 0 ? 'pointer' : '')};
-
-  > div {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-    align-items: center;
-    justify-content: center;
   }
 `;
