+++ src/services/activityApi.js
@@ -10,6 +10,16 @@ export async function getActivities(token) {
   return response.data;
 }
 
+export async function getNumberOfEnrollmentsByActivity(token, activityId) {
+  const response = await api.get(`/activities/activity-tickets/${activityId}`, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    }
+  });
+
+  return response.data;
+}
+
 export async function create(body, token) {
   const response = await api.post('/activities', body, {
     headers: {
